refactor(view): extract cursor, insert, entry_format modules
Extract three focused modules from view.lua (1442 -> 1006 lines, -30%):

- cursor.lua (100 lines): cursor position tracking and constraint logic
  (set_last_cursor, maybe_set_cursor, get_last_cursor, constrain_cursor,
  calc_constrained_cursor_pos)
- insert.lua (158 lines): insert-mode cursor constraints and visual guide
  (show_insert_guide, update_insert_boundary, setup_insert_constraints,
  clear_boundary)
- entry_format.lua (227 lines): entry line rendering and highlight
  computation (format_entry_line, compute_highlights_for_cols,
  is_unix_executable, get_link_text)

get_col_pad promoted to M.get_col_pad on view.lua; new modules access it
via lazy require to avoid circular dependencies. Decoration provider stays
in view.lua due to tight coupling with session state. Thin re-export
wrappers on M preserve backwards compatibility for external callers.
